pyproject.toml
[project]
name = "oidc_pure"
version = "0.1.0"
description = "Pure OAuth2/OIDC implementation following RFC 6749 for Keycloak and multiple identity providers"
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.12"
authors = [
    { name = "Paulo Alexandre" }
]
maintainers = [
    { name = "Paulo Alexandre" }
]
keywords = [
    "oauth2",
    "oidc",
    "openid-connect",
    "keycloak",
    "authentication",
    "authorization",
    "identity",
    "rfc6749",
    "pkce",
    "jwt",
    "google-oauth",
    "github-oauth",
    "microsoft-oauth",
    "multi-provider",
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Programming Language :: Python :: 3.14",
    "Programming Language :: Python :: Implementation :: CPython",
    "Topic :: Security",
    "Topic :: Security :: Cryptography",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: Internet :: WWW/HTTP",
    "Framework :: Django",
    "Framework :: FastAPI",
    "Framework :: Flask",
    "Typing :: Typed",
    "Operating System :: OS Independent",
]

dependencies = [
    "httpx>=0.27.0",
    "python-dotenv>=1.2.1",
]

[project.urls]
Homepage = "https://github.com/Pacalexandre/oidc_pure"
Repository = "https://github.com/Pacalexandre/oidc_pure"
Documentation = "https://github.com/Pacalexandre/oidc_pure#readme"
"Bug Tracker" = "https://github.com/Pacalexandre/oidc_pure/issues"
Security = "https://github.com/Pacalexandre/oidc_pure/blob/main/SECURITY.md"

[project.optional-dependencies]
dev = [
    "ruff>=0.3.0",
    "pip-audit>=2.10.0",
    "safety>=3.7.0",
    "bandit>=1.9.3",
]
test = [
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "pytest-cov>=7.0.0",
    "pytest-mock>=3.15.1",
    "respx>=0.22.0",
]
django = [
    "django>=4.2",
]
drf = [
    "djangorestframework>=3.14",
    "drf-spectacular>=0.26.0",
]
fastapi = [
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0",
]
flask = [
    "flask>=3.0.0",
]
all = [
    "oidc_pure[dev,test,django,drf,fastapi,flask]",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["oidc_pure"]

[tool.hatch.build.targets.sdist]
include = [
    "/oidc_pure",
    "/tests",
    "/examples",
    "README.md",
    "LICENSE",
    "SECURITY.md",
    "pyproject.toml",
]
exclude = [
    "**/__pycache__",
    "**/*.pyc",
    "**/.pytest_cache",
    "**/htmlcov",
    "**/.coverage",
    ".env",
    ".env.*",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "-v",
    "--strict-markers",
    "--strict-config",
    "--cov=oidc_pure",
    "--cov-report=term-missing",
    "--cov-report=html",
]

[tool.coverage.run]
source = ["oidc_pure"]
omit = [
    "tests/*",
    "*/test_*.py",
    "*/__pycache__/*",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W"]
ignore = ["E501"]  # line too long (handled by formatter)
